# dsp-backend/openapi.yaml
openapi: 3.1.0
info:
  title: DSP (Digital Security Platform) API
  description: |
    Backend API for the Digital Security Platform — a TARA (Threat Analysis and Risk Assessment)
    pipeline that ingests system artifacts, identifies threats, matches CVEs, builds attack paths,
    scores risk, and generates mitigations.

    ## Pipeline Overview
    1. **Ingest** — parse uploaded artifacts into a canonical system model
    2. **Threats** — STRIDE + HEAVENS threat identification
    3. **CVE Matching** — match software instances to known vulnerabilities (3-tier: exact CPE → virtual CPE → keyword)
    4. **Attack Paths** — construct feasible attack paths across trust boundaries via BFS + LLM plausibility
    5. **Risk Scoring** — score each finding by likelihood, impact, exploitability, and exposure
    6. **Mitigations** — generate control recommendations per risk item
    7. **Export** — produce reports in JSON, Markdown, or PDF

    ## Authentication
    All protected endpoints require a valid JWT sent as an **httpOnly cookie** named `token`.
    Obtain a token via `POST /auth/sign-up` or `POST /auth/sign-in`. Tokens expire after **7 days**.

    ## Response Envelope
    All responses follow a consistent envelope:
    ```json
    { "ok": true,  "data": <resource> }
    { "ok": false, "error": { "code": "...", "message": "..." } }
    ```

    ## Async Operations
    Run creation (`POST /projects/:projectId/runs`) and export creation
...
  version: 1.1.0
  contact:
...
servers:
...
paths:
...
  /auth/profile:
    get:
      operationId: getProfile
      summary: Get the authenticated user's profile
      description: Returns the full profile of the currently authenticated user.
      tags:
        - Auth
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Profile returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ---------------------------------------------------------------------------
  # Users
  # ---------------------------------------------------------------------------
  /users/me/active-project:
    patch:
      operationId: setActiveProject
      summary: Set or clear the active project
      description: |
        Stores the user's currently active project ID for UI state persistence.
        Pass `projectId: null` to clear the active project.
        The project must be owned by the requesting user.
      tags:
        - Users
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetActiveProjectRequest"
            examples:
              set:
                summary: Set active project
                value:
                  projectId: cma1b2c3d4e5f6g7h8i9j
              clear:
                summary: Clear active project
                value:
                  projectId: null
      responses:
        "200":
          description: Active project updated — returns updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Project not found or not owned by requesting user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                ok: false
                error:
                  code: NOT_FOUND
                  message: Project not found or not owned by you
  /users/{userId}:
    get:
      operationId: getUser
      summary: Get a user by ID
      tags:
        - Users
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Projects
  # ---------------------------------------------------------------------------
  /projects:
    get:
      operationId: listProjects
      summary: List the authenticated user's projects
      description: Returns all projects created by the authenticated user, sorted by creation date descending.
      tags:
        - Projects
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createProject
      summary: Create a project
      description: Creates a new project owned by the authenticated user.
      tags:
        - Projects
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
            example:
              name: Vehicle Gateway Analysis
              domain: automotive
              systemContext: Central gateway ECU bridging OBD-II, CAN bus, and Ethernet backbone.
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /projects/{projectId}:
    get:
      operationId: getProject
      summary: Get a project by ID
      description: |
        Returns full project details including creator info and the last 10 runs.
        Only the project owner can access this endpoint.
      tags:
        - Projects
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      operationId: updateProject
      summary: Update project system context
      description: |
        Updates the project's system context description.
        Only the project owner can update.
      tags:
        - Projects
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
            example:
              systemContext: Updated description of the system under analysis.
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Runs
  # ---------------------------------------------------------------------------
  /projects/{projectId}/runs:
    get:
      operationId: listRunsByProject
      summary: List runs for a project
      description: |
        Returns all runs for the project, sorted by `startedAt` descending.
        Only the project owner can list runs.
      tags:
        - Runs
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Run list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      operationId: createRun
      summary: Start an analysis run
      description: |
        Creates a new run and queues it for asynchronous pipeline processing.
        Returns **202 Accepted** immediately. Poll `GET /runs/{runId}` to track status.

        At least one artifact (type `text`) must be provided. Artifacts are parsed
        during the ingestion step to build the canonical system model.
      tags:
        - Runs
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunRequest"
            example:
              artifacts:
                - type: text
                  content: "System: Central Gateway ECU\nAssets: OBD port, CAN bus, Ethernet..."
      responses:
        "202":
          description: Run queued for processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /runs/{runId}/software-instances:
    get:
      operationId: getSoftwareInstances
      summary: List software instances for a run
      description: |
        Returns all software instances extracted during ingestion for this run.
        Independent of CVE matching — components with 0 CVEs are included.
      tags:
        - Runs
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Software instance list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoftwareInstanceListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # CVEs
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: |
        JWT stored in an httpOnly cookie named `token`. Obtain via
        `POST /auth/sign-up` or `POST /auth/sign-in`. Expires after 7 days.

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      description: CUID of the user
      schema:
        type: string
        minLength: 1
    ProjectId:
      name: projectId
      in: path
      required: true
      description: CUID of the project
      schema:
        type: string
        minLength: 1
    RunId:
      name: runId
      in: path
      required: true
      description: CUID of the run
      schema:
        type: string
        minLength: 1

  # ---------------------------------------------------------------------------
  # Reusable Responses
  # ---------------------------------------------------------------------------
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            ok: false
            error:
              code: VALIDATION_ERROR
              message: "Invalid request: email must be a valid email address"
    Unauthorized:
      description: Authentication required or token invalid/expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            ok: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
    Forbidden:
      description: Authenticated but not allowed to access this resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            ok: false
            error:
              code: FORBIDDEN
              message: Access denied
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            ok: false
            error:
              code: NOT_FOUND
              message: Resource not found.
...
  schemas:
...
    UserRole:
      type: string
      enum: [owner, analyst, reviewer, viewer]
      description: User's role within the platform
    ProjectDomain:
      type: string
      enum: [automotive, general_system_security]
      description: Domain context for the project
    ProjectStatus:
      type: string
      enum: [active, archived]
    RunStatus:
      type: string
      enum: [queued, running, completed, failed]
      description: Current status of the analysis pipeline run
    StepStatus:
      type: string
      enum: [pending, running, completed, failed]
    ArtifactType:
      type: string
      enum: [text]
      description: Type of input artifact
...
    ErrorEnvelope:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error description

    # ---- Health ----
...
    User:
      type: object
      required: [id, name, email, role, createdAt, lastActiveAt]
      properties:
        id:
          type: string
          description: CUID identifier
        name:
          type: string
          minLength: 1
          maxLength: 255
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/UserRole"
        activeProjectId:
          type: ["string", "null"]
          description: CUID of the user's currently active project; null if not set
        createdAt:
          type: string
          format: date-time
        lastActiveAt:
          type: string
          format: date-time
    SetActiveProjectRequest:
      type: object
      required: [projectId]
      properties:
        projectId:
          type: ["string", "null"]
          description: CUID of the project to activate, or null to clear
    UserResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/User"

    # ---- Project ----
    Project:
      type: object
      required: [id, name, domain, status, createdBy, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        domain:
          $ref: "#/components/schemas/ProjectDomain"
        status:
          $ref: "#/components/schemas/ProjectStatus"
        systemContext:
          type: ["string", "null"]
        createdBy:
          type: string
          description: User ID of the project owner
        createdAt:
          type: string
          format: date-time
        creator:
          type: object
          description: Populated when fetching a project by ID
          required: [id, name]
          properties:
            id:
              type: string
            name:
              type: string
        runs:
          type: array
          description: Last 10 runs — populated when fetching by ID
          items:
            type: object
            required: [id, status, startedAt]
            properties:
              id:
                type: string
              status:
                $ref: "#/components/schemas/RunStatus"
              startedAt:
                type: string
                format: date-time
              completedAt:
                type: ["string", "null"]
                format: date-time
    CreateProjectRequest:
      type: object
      required: [name, domain]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        domain:
          $ref: "#/components/schemas/ProjectDomain"
        systemContext:
          type: string
          description: Free-text description of the system under analysis
    UpdateProjectRequest:
      type: object
      required: [systemContext]
      properties:
        systemContext:
          type: string
          minLength: 1
          description: Updated system context description
    ProjectResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/Project"
    ProjectListResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          type: array
          items:
            $ref: "#/components/schemas/Project"

    # ---- Run ----
    RunSteps:
      type: object
      description: Per-step status of the analysis pipeline
      required: [ingestion, threats, cves, attack_paths, risk, mitigations]
      properties:
        ingestion:
          $ref: "#/components/schemas/StepStatus"
        threats:
          $ref: "#/components/schemas/StepStatus"
        cves:
          $ref: "#/components/schemas/StepStatus"
        attack_paths:
          $ref: "#/components/schemas/StepStatus"
        risk:
          $ref: "#/components/schemas/StepStatus"
        mitigations:
          $ref: "#/components/schemas/StepStatus"
    Run:
      type: object
      required: [id, projectId, initiatedBy, status, steps, startedAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        initiatedBy:
          type: string
          description: User ID who initiated the run
        status:
          $ref: "#/components/schemas/RunStatus"
        steps:
          $ref: "#/components/schemas/RunSteps"
        errorMessage:
          type: ["string", "null"]
        failedStep:
          type: ["string", "null"]
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        modelVersion:
          type: ["string", "null"]
        retrievalIndexVersion:
          type: ["string", "null"]
        configSnapshot:
          type: ["object", "null"]
          additionalProperties: true
    InputArtifact:
      type: object
      required: [type, content]
      properties:
        type:
          $ref: "#/components/schemas/ArtifactType"
        content:
          type: string
          minLength: 1
          description: Raw text content of the artifact
    CreateRunRequest:
      type: object
      required: [artifacts]
      properties:
        artifacts:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/InputArtifact"
        configSnapshot:
          type: object
          additionalProperties: true
          description: Optional configuration overrides for this run
    RunResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/Run"
    RunListResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          type: array
          items:
            $ref: "#/components/schemas/Run"

    # ---- Threat ----
...
    SoftwareInstance:
      type: object
      required: [id, name, assetId, assetName, assetKind]
      properties:
        id:
          type: string
        assetId:
          type: string
        assetName:
          type: string
        assetKind:
          type: string
        name:
          type: string
        version:
          type: ["string", "null"]
        cpe:
          type: ["string", "null"]
    SoftwareInstanceListResponse:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          type: array
          items:
            $ref: "#/components/schemas/SoftwareInstance"
...